Extract out OperationGenerator
